pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "otto"
version = "0.1.0"
description = "AI voice controller for model railways — Rocrail layout control as MCP tools for Claude Desktop"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
    "thefuzz>=0.22.0",
]

[project.optional-dependencies]
voice = [
    "faster-whisper",
    "kokoro-onnx",
    "sounddevice",
    "soundfile",
    "pyautogui",
    "pynput",
    "numpy",
    "pyperclip",
]

[project.scripts]
otto = "otto.mcp_server:main"
otto-voice = "otto.voice_daemon:main"
